This PR introduces initial custom class bindings for ProcessGroup, and
make it work specifically for ProcessGroupNCCL.

We didn't make the API exactly the same with python pybind bindings, as
there are some issues with TorchBind to support some argument types and
we will resolve that separately to make these two stays the same. And
eventually, when we have a good TorchBind coverage, we will switch all
bindings in c10d TorchBind

Differential Revision: [D23228682](https://our.internmc.facebook.com/intern/diff/D23228682)

[ghstack-poisoned]

+++ torch/lib/c10d/ProcessGroup.cpp
@@ -1,8 +1,6 @@
 #include <c10d/ProcessGroup.hpp>
 
 #include <c10/util/Logging.h>
-#include <sys/types.h>
-#include <torch/custom_class.h>
 
 namespace c10d {
 
@@ -79,7 +77,7 @@ std::exception_ptr ProcessGroup::Work::exception() const {
   return exception_;
 }
 
-int64_t ProcessGroup::Work::sourceRank() const {
+int ProcessGroup::Work::sourceRank() const {
   throw std::runtime_error(
       "sourceRank() may only be called on work objects "
       "that correspond to a recv or recv-from-any call.");
